refactor: segment app storage api boundary
