+++ rustfs/src/app/admin_usecase.rs
@@ -14,13 +14,13 @@
 
 //! Admin application use-case contracts.
 
-use super::storage_api::StorageAdminApi;
-use super::storage_api::admin::get_server_info;
-use super::storage_api::capacity::{
+use super::storage_api::admin_usecase::StorageAdminApi;
+use super::storage_api::admin_usecase::admin::get_server_info;
+use super::storage_api::admin_usecase::capacity::{
     PoolDecommissionInfo, PoolStatus, RebalStatus, get_total_usable_capacity, get_total_usable_capacity_free,
 };
-use super::storage_api::data_usage::{apply_bucket_usage_memory_overlay, load_data_usage_from_backend};
-use super::storage_api::{ECStore, EndpointServerPools};
+use super::storage_api::admin_usecase::data_usage::{apply_bucket_usage_memory_overlay, load_data_usage_from_backend};
+use super::storage_api::admin_usecase::{ECStore, EndpointServerPools};
 use crate::app::runtime_sources::{
     AppContext, current_app_context, resolve_endpoints_handle, resolve_object_store_handle_for_context,
 };
@@ -597,7 +597,7 @@ impl DefaultAdminUsecase {
 
 #[cfg(test)]
 mod tests {
-    use super::super::storage_api::capacity::{PoolDecommissionInfo, PoolStatus};
+    use super::super::storage_api::admin_usecase::capacity::{PoolDecommissionInfo, PoolStatus};
     use super::*;
     use time::OffsetDateTime;
