+++ rustfs/src/app/select_object.rs
@@ -1,7 +1,7 @@
-use super::storage_api::ObjectOperations as _;
-use super::storage_api::options::get_opts;
-use super::storage_api::request_context::spawn_traced;
-use super::storage_api::{get_validated_store, validate_sse_headers_for_read, validate_ssec_for_read};
+use super::storage_api::select_object::ObjectOperations as _;
+use super::storage_api::select_object::options::get_opts;
+use super::storage_api::select_object::request_context::spawn_traced;
+use super::storage_api::select_object::{get_validated_store, validate_sse_headers_for_read, validate_ssec_for_read};
 use crate::app::runtime_sources::resolve_s3select_db;
 use crate::error::ApiError;
 use bytes::Bytes;
